tg_bot/pyproject.toml
[tool.poetry]
name = "tg-bot"
version = "0.1.0"
description = ""
authors = ["Alisher <alisherxujanov163@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.10"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
